+++ src/App.js
@@ -1,14 +1,13 @@
 import React from "react";
-import "./App.scss";
-import { AboutMe } from "./pages/AboutMe";
-import { Header } from "./components/Header";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import { Home } from "./pages/Home";
-import { Projects } from "./pages/Projects";
-
-import { MyVideos } from "./pages/MyVideos";
-import { MyPictures } from "./pages/MyPictures";
-import { Resume } from "./resume/src/resume";
+import "./App.scss";
+import { AboutMe } from "./components/aboutMe/aboutMe";
+import { Header } from "./components/header/header";
+import { Home } from "./components/home/home";
+import { Projects } from "./components/projects/projects";
+import { MyVideos } from "./components/myVideos/myVideos";
+import { MyPictures } from "./components/myPictures/myPictures";
+import { Resume } from "./components/resume/resume";
 
 function App() {
   return (
