+++ src/components/home/home.js
@@ -1,0 +1,13 @@
+import React from "react";
+export const Home = () => {
+  return (
+    <div className="page-info">
+      <h3>Home</h3>
+      <p>
+        Welome to my page. Thanks for stopping by. This is a work in progress
+        that I will constantly be updating. Feel free to take a look around, and
+        if you have any suggestions inbox me on the contact page.
+      </p>
+    </div>
+  );
+};
