+++ src/components/myPictures/myPictures.js
