+++ src/components/resume/resume.scss
@@ -229,6 +229,9 @@ a:focus {
   .header {
     display: none;
   }
+  nav .navMenu {
+    display: none;
+  }
 
   .app {
     padding-bottom: 0;
