+++ src/objects/nav.json
@@ -4,5 +4,6 @@
   { "title": "Projects", "link": "/Projects" },
   { "title": "My Resume", "link": "/myResume" },
   { "title": "My Videos", "link": "/myVideos" },
-  { "title": "My Pictures", "link": "/myPictures" }
+  { "title": "My Pictures", "link": "/myPictures" },
+  { "title": "Contact", "link": "/contact" }
 ]
