--- src/pages/Home.js
@@ -1,9 +1,0 @@
-import React from "react";
-export const Home = () => {
-  return (
-    <div className="page-info">
-      <h3>Home</h3>
-      <p>Welome to my page. Thanks for stopping by. </p>
-    </div>
-  );
-};
